% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{connect_to_wordbank}
\alias{connect_to_wordbank}
\title{Connect to the Wordbank database}
\usage{
connect_to_wordbank(mode = "prod")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
\code{"prod"}, or \code{"dev"} (defaults to \code{"prod"})}
}
\value{
A \code{src} object which is connection to the Wordbank database
}
\description{
Connect to the Wordbank database
}
\examples{
wordbank <- connect_to_wordbank()
}

