% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{find_database}
\alias{find_database}
\title{Locate Wordbank Database}
\usage{
find_database(mode = "remote")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
\code{"remote"} (defaults to \code{"remote"}), or a date given in one
of the following formats: \code{"YYYY-MM-DD"}, \code{"YYYY/MM/DD"}, 
\code{"MM-DD-YYYY"}, or \code{"MM/DD/YYYY"}.}
}
\value{
A \code{src} object which is a connection to the Wordbank database.
}
\description{
Locate Wordbank Database
}
\examples{
\dontrun{
db_info <- find_database("2017-03-22")
}
}
\keyword{internal}
