% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{find_matches}
\alias{find_matches}
\title{Connect to the Wordbank database}
\usage{
find_matches(x, mode = "remote")
}
\arguments{
\item{x}{A 1-row dataframe passed on from \code{"match_crossling_items"} 
with the following variables:}

\item{mode}{A string indicating connection mode: one of \code{"local"},
or \code{"remote"} (defaults to \code{"remote"})}
}
\value{
A dataframe describing the instrument (\code{language}), its variables (\code{item_id}, 
  \code{definition}, \code{uni_lemma}, \code{lexical_category}, \code{lexical_class}),
  its agegroup (\code{age}, \code{n_children}), and the group's performance (\code{comprehension}, 
  \code{production}, \code{comprehension_sd}, \code{production_sd}).
}
\description{
Connect to the Wordbank database
}
\examples{
\dontrun{
dog_unilemma <- get_item_data(language = "Italian", form = "WG") \%>\%
  filter(uni_lemma == "dog")
dog_in_italian <- find_matches(dog_unilemma)
}
}
\keyword{internal}
