% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_common_table}
\alias{get_common_table}
\title{Connect to a single Wordbank common table}
\usage{
get_common_table(src, name)
}
\arguments{
\item{src}{A connection to the Wordbank database}

\item{name}{A string indicating the name of a common table}
}
\value{
A \code{tbl} object
}
\description{
Connect to a single Wordbank common table
}
\examples{
\dontrun{
src <- connect_to_wordbank()
instruments <- get_common_table(src, "instrument")
rm(src, instruments)
}
}
\keyword{internal}

