% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_common_tables}
\alias{get_common_tables}
\title{Connect to all the Wordbank common tables}
\usage{
get_common_tables(mode = "prod")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
\code{"prod"}, or \code{"dev"} (defaults to \code{"prod"})}
}
\value{
A list whose names are common table names and whose values
are \code{tbl} objects
}
\description{
Connect to all the Wordbank common tables
}
\examples{
common_tables <- get_common_tables()
}

