% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_instrument_table}
\alias{get_instrument_table}
\title{Connect to an instrument's Wordbank table}
\usage{
get_instrument_table(language, form, mode = "prod")
}
\arguments{
\item{language}{A string of the instrument's language (insensitive to case
and whitespace)}

\item{form}{A string of the instrument's form (insensitive to case and
whitespace)}

\item{mode}{A string indicating connection mode: one of \code{"local"},
\code{"prod"}, or \code{"dev"} (defaults to \code{"prod"})}
}
\value{
A \code{tbl} object containing the instrument's data
}
\description{
Connect to an instrument's Wordbank table
}
\examples{
eng_ws <- get_instrument_table("english", "ws")
}

