% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_wordbank_args}
\alias{get_wordbank_args}
\title{Get database connection arguments}
\usage{
get_wordbank_args()
}
\value{
List of database connection arguments: host, db_name, username,
  password
}
\description{
Get database connection arguments
}
\examples{
\donttest{
get_wordbank_args()
}
}
