% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{match_crossling_items}
\alias{match_crossling_items}
\title{Match unilemmas to their Wordbank group-by-item data}
\usage{
match_crossling_items(unilemmas, mode = "remote")
}
\arguments{
\item{unilemmas}{A character vector of unilemmas to get group-by-item data for}

\item{mode}{A string indicating connection mode: one of \code{"local"},
or \code{"remote"} (defaults to \code{"remote"})}
}
\value{
A data frame where each row is a particular CDI item with a set of
variables describing its instrument (\code{language}), its variables (\code{item_id},
 \code{definition}, \code{uni_lemma}, \code{lexical_category}, \code{lexical_class}),
 its agegroup (\code{age}, \code{n_children}), and the group's performance (\code{comprehension},
 \code{production}, \code{comprehension_sd}, \code{production_sd}).
}
\description{
Match unilemmas to their Wordbank group-by-item data
}
\examples{
\dontrun{
crossling_words <- match_crossling_items(unilemmas = c("hat", "nose"))
}
}
