% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{connect_to_wordbank}
\alias{connect_to_wordbank}
\title{Connect to the Wordbank database}
\usage{
connect_to_wordbank(mode = "remote")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
or \code{"remote"} (defaults to \code{"remote"})}
}
\value{
A \code{src} object which is connection to the Wordbank database
}
\description{
Connect to the Wordbank database
}
\examples{
\dontrun{
wordbank <- connect_to_wordbank()
rm(wordbank)
}
}
\keyword{internal}

