% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{connect_to_wordbank}
\alias{connect_to_wordbank}
\title{Connect to the Wordbank database}
\usage{
connect_to_wordbank(db_args = NULL)
}
\arguments{
\item{db_args}{List with arguments to connect to wordbank mysql database
(host, dbname, user, and password).}
}
\value{
A \code{src} object which is connection to the Wordbank database.
}
\description{
Connect to the Wordbank database
}
\examples{
\donttest{
src <- connect_to_wordbank()
}
}
