% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_common_tables}
\alias{get_common_tables}
\title{Connect to all the Wordbank common tables}
\usage{
get_common_tables(src)
}
\arguments{
\item{src}{A connection to the Wordbank database}
}
\value{
A list whose names are common table names and whose values
are \code{tbl} objects
}
\description{
Connect to all the Wordbank common tables
}
\examples{
\dontrun{
src <- connect_to_wordbank()
common_tables <- get_common_tables(src)
RMySQL::dbDisconnect(src$con)
}
}

