% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossling.R
\name{get_crossling_items}
\alias{get_crossling_items}
\title{Get the uni_lemmas available in Wordbank}
\usage{
get_crossling_items(db_args = NULL)
}
\arguments{
\item{db_args}{Deprecated, ignored.}
}
\value{
A data frame with the column \code{uni_lemma}.
}
\description{
Get the uni_lemmas available in Wordbank
}
\examples{
\dontrun{
uni_lemmas <- get_crossling_items()
}
}
