% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_crossling_items}
\alias{get_crossling_items}
\title{Get a list of unilemmas available at Wordbank}
\usage{
get_crossling_items(mode = "remote")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
or \code{"remote"} (defaults to \code{"remote"})}
}
\value{
A data frame
}
\description{
Get a list of unilemmas available at Wordbank
}
\examples{
\dontrun{
unilemmas <- get_crossling_items()
}
}
