% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossling.R
\name{get_crossling_items}
\alias{get_crossling_items}
\title{Get the uni_lemmas available in Wordbank}
\usage{
get_crossling_items(mode = "remote")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
\code{"remote"} (defaults to \code{"remote"}), or a date given in one
of the following formats: \code{"YYYY-MM-DD"}, \code{"YYYY/MM/DD"}, 
\code{"MM-DD-YYYY"}, or \code{"MM/DD/YYYY"}.}
}
\value{
A data frame with the column \code{uni_lemma}.
}
\description{
Get the uni_lemmas available in Wordbank
}
\examples{
\dontrun{
uni_lemmas <- get_crossling_items()
}
}
