% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossling.R
\name{get_crossling_items}
\alias{get_crossling_items}
\title{Get the unilemmas available in Wordbank}
\usage{
get_crossling_items(mode = "remote")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"}, or
\code{"remote"} (defaults to \code{"remote"}).}
}
\value{
A data frame with the column \code{uni_lemma}.
}
\description{
Get the unilemmas available in Wordbank
}
\examples{
\dontrun{
unilemmas <- get_crossling_items()
}
}
