% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_datasets}
\alias{get_datasets}
\title{Get the Wordbank data sources}
\usage{
get_datasets(language = NULL, form = NULL, admin_data = FALSE, db_args = NULL)
}
\arguments{
\item{language}{An optional string specifying which language's datasets to
retrieve.}

\item{form}{An optional string specifying which form's datasets to
retrieve.}

\item{admin_data}{A logical indicating whether to include the number of
administrations in the dataset.}

\item{db_args}{Deprecated, ignored.}
}
\value{
A data frame where each row is a particular dataset and its
  characteristics.
}
\description{
Get the Wordbank data sources
}
\examples{
\dontrun{
english_ws_datasets <- get_datasets("English (American)", "WS")
}
}
