% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_instrument_data}
\alias{get_instrument_data}
\title{Get the Wordbank administration-by-item data}
\usage{
get_instrument_data(instrument_table, items)
}
\arguments{
\item{instrument_table}{A Wordbank instrument table as returned by
\code{get_instrument_table}}

\item{items}{A character vector of column names of \code{instrument_table} of
items to extract. If not supplied, defaults to all the columns of
\code{instrument_table}}
}
\value{
A data frame where each row is the result (\code{value}) of a given item
  (\code{num_item_id}) for a given administration (\code{data_id})
}
\description{
Get the Wordbank administration-by-item data
}
\examples{
eng_ws <- get_instrument_table("English", "WS")
eng_ws_data <- get_instrument_data(eng_ws, c("item_1", "item_42"))
}

