% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_instruments}
\alias{get_instruments}
\title{Get the Wordbank instruments}
\usage{
get_instruments(db_args = NULL)
}
\arguments{
\item{db_args}{Deprecated, ignored.}
}
\value{
A data frame where each row is a CDI instrument and each column is
  a variable about the instrument (\code{instrument_id}, \code{language},
  \code{form}, \code{form_type}, \code{age_min}, \code{age_max},
  \code{has_grammar}, \code{unilemma_coverage}).
}
\description{
Get the Wordbank instruments
}
\examples{
\dontrun{
instruments <- get_instruments()
}
}
