% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr.R
\name{get_instruments}
\alias{get_instruments}
\title{Get the Wordbank instruments}
\usage{
get_instruments(mode = "remote")
}
\arguments{
\item{mode}{A string indicating connection mode: one of \code{"local"},
or \code{"remote"} (defaults to \code{"remote"})}
}
\value{
A data frame
}
\description{
Get the Wordbank instruments
}
\examples{
\dontrun{
instruments <- get_instruments()
}
}
\keyword{internal}

