% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr-package.R
\docType{package}
\name{wordbankr-package}
\alias{wordbankr}
\alias{wordbankr-package}
\title{wordbankr: Accessing the 'Wordbank' Database}
\description{
Connecting to 'Wordbank' \url{https://wordbank.stanford.edu}, an open repository for developmental vocabulary data from the MacArthur-Bates Communicative Development Inventories (Frank et al. 2017 \doi{10.1017/S0305000916000209}). Data are read from a versioned dataset hosted on 'Redivis', so analyses can be pinned to a specific release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://langcog.github.io/wordbankr/}
  \item \url{https://github.com/langcog/wordbankr/}
  \item Report bugs at \url{https://github.com/langcog/wordbankr/issues/}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Other contributors:
\itemize{
  \item Daniel Yurovsky \email{yurovsky@stanford.edu} [contributor]
  \item Michael Frank \email{mcfrank@stanford.edu} [contributor]
  \item Danielle Kellier \email{dkellier@stanford.edu} [contributor]
  \item Alvin Tan \email{tanawm@stanford.edu} [contributor]
}

}
\keyword{internal}
