% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordbankr-package.R
\docType{package}
\name{wordbankr-package}
\alias{wordbankr}
\alias{wordbankr-package}
\title{wordbankr: Accessing the Wordbank Database}
\description{
Connecting to Wordbank, an open repository for developmental vocabulary data. For more information on the underlying data, see \url{https://wordbank.stanford.edu}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://langcog.github.io/wordbankr/}
  \item \url{https://github.com/langcog/wordbankr/}
  \item Report bugs at \url{https://github.com/langcog/wordbankr/issues/}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
}

Other contributors:
\itemize{
  \item Daniel Yurovsky \email{yurovsky@stanford.edu} [contributor]
  \item Michael Frank \email{mcfrank@stanford.edu} [contributor]
  \item Danielle Kellier \email{dkellier@stanford.edu} [contributor]
  \item Alvin Tan \email{tanawm@stanford.edu} [contributor]
}

}
\keyword{internal}
